[feat] [rename] JWT 토큰으로 스케쥴 정보 가져오는 기능 개발 및 리네이밍
