+++ src/main/java/Ness/Backend/domain/schedule/ScheduleService.java
@@ -2,12 +2,10 @@ package Ness.Backend.domain.schedule;
 
 import Ness.Backend.domain.schedule.dto.ScheduleDto;
 import Ness.Backend.domain.schedule.dto.ScheduleListResponseDto;
-import Ness.Backend.domain.entity.Member;
-import Ness.Backend.domain.entity.Schedule;
-import Ness.Backend.domain.entity.ScheduleDate;
+import Ness.Backend.domain.member.entity.Member;
+import Ness.Backend.domain.schedule.entity.Schedule;
 import Ness.Backend.domain.member.MemberRepository;
 import Ness.Backend.domain.schedule.dto.ScheduleCreateRequestDto;
-import Ness.Backend.domain.schedule.dto.ScheduleDateDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -26,34 +24,35 @@ public class ScheduleService {
 
         // ScheduleListResponseDto에 매핑
         List<ScheduleDto> scheduleDtos = scheduleList.stream()
-                .map(schedule -> new ScheduleDto.ScheduleDtoBuilder()
+                .map(schedule -> ScheduleDto.builder()
                         .id(schedule.getId())
                         .info(schedule.getInfo())
+                        .date(schedule.getDate())
                         .location(schedule.getLocation())
-                        .scheduleDateDto(new ScheduleDateDto.ScheduleDateDtoBuilder()
-                                .time(schedule.getScheduleDate().getTime())
-                                .date(schedule.getScheduleDate().getDate())
-                                .build())
                         .build())
                 .toList();
         return new ScheduleListResponseDto(scheduleDtos);
     }
 
     @Transactional
-    public Long createNewSchedule(ScheduleCreateRequestDto scheduleCreateRequestDto){
-        Member memberEntity = memberRepository.findMemberById(scheduleCreateRequestDto.getMember_id());
+    public Long createNewSchedule(Long id, ScheduleCreateRequestDto scheduleCreateRequestDto){
+        Member memberEntity = memberRepository.findMemberById(id);
 
+        /*
         ScheduleDate newScheduleDate = ScheduleDate.builder()
                 .time(scheduleCreateRequestDto.getScheduleDateDto().getTime())
                 .date(scheduleCreateRequestDto.getScheduleDateDto().getDate())
                 .build();
+
+         */
 
         //새로운 채팅 생성
         Schedule newSchedule = Schedule.builder()
                 .info(scheduleCreateRequestDto.getInfo())
                 .location(scheduleCreateRequestDto.getLocation())
                 .person(scheduleCreateRequestDto.getPerson())
-                .scheduleDate(newScheduleDate)
+                .date(scheduleCreateRequestDto.getDate())
+                //.date(LocalDateTime.now(ZoneId.of("Asia/Seoul")).atZone(ZoneId.of("Asia/Seoul")))
                 .member(memberEntity)
                 //.category() //이 연관관계들은 나중에 넣어야 함
                 //.chat()
